+++ src/storage/plastic_db.h
@@ -7,12 +7,12 @@ class PlasticDB {
  public:
   explicit PlasticDB(const QString &path);
   bool init();
-  bool addPlastic(const QString &name, double weight, double cost);
+  bool addPlastic(const QString &name, int weight, double cost);
   QHash<QString, QVariant> getPlasticByName(const QString &name);
   bool deletePlasticByName(const QString &name);
   QList<QString> getPlasticList();
   bool updatePlasticByName(const QString &oldName, const QString &newName,
-                           double weight, double cost);
+                           int weight, double cost);
 
  private:
   QSqlDatabase db;
