+++ src/test/reference/calculating.h
@@ -5,9 +5,9 @@
 #include <stdlib.h>
 
 // Объявления функций
-double cost_prise(double p, double t, double h, double md, double d, double st,
-                  double mk, double am, double post, int x);
+double cost_prise(int p, int t, double h, double md, double d, double st,
+                  int mk, double am, double post, int x);
 double calculating(double cost, double mod, double marg);
-double amortization(double a, double t, double spi);
+double amortization(double a, int t, double spi);
 
 #endif  // CALCULATING_H
